move projen specific property out of model
